Added service project and updated app
Lots of new changes and further improvements to the app and the service.
The service is being built into Azure Web Site.
Adding support for routes in the app, as oppose to all views embedded
into the index.html.
Added grunt to launch a live reload for development purposes.
This code base is not really a fully functional codebase, so please
don't use it for anything other than looking at the source.