Change theme to Cyan
Changed the theme to Cyan.
Added a preload screen that will be displayed when loading tasks is
running, such as finding out which OS the user is running, which will
affect the window control icons.
The color fading wallpaper is included in this commit, but will be
removed moving forward with a regular white (or photo wallpaper)
background.