Improve service
This commit contains plenty of improvements to the service. The service
no longer creates a DocumentDB instance for each generated login URL.
Instead if relies on in-memory array that will purge items that are
older than 30 minutes. This avoids cluttering the DocumentDB and
simplifies code.

Some grunt packages have been updated. The others have been carefully
selected NOT to be upgraded, as they create issues during compilation of
app.